@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using quiz_management_system.App.Helpers;
 using quiz_management_system.Application.Features.GoogleLogin;
+using quiz_management_system.Application.Features.UsersFeatures.Queries.GetMe;
 using quiz_management_system.Contracts.Reponses.Identity;
 using quiz_management_system.Domain.Common.ResultPattern.Result;
 using quiz_management_system.Infrastructure.Idenitity;
@@ -17,72 +18,105 @@ namespace quiz_management_system.App.Controllers;
 [Tags("Identity - External Auth")]
 [Produces("application/json")]
 [ApiVersion("1.0")]
-public sealed class ExternalAuthController : ControllerBase
+public sealed class ExternalAuthController(
+    IAuthCookieWriter authCookieWriter,
+    ISender sender,
+    SignInManager<ApplicationUser> signInManager, FrontendOptions frontendOptions
+) : ControllerBase
 {
     private const string GoogleProvider = "Google";
 
-    private readonly ISender sender;
-    private readonly SignInManager<ApplicationUser> signInManager;
-
-    public ExternalAuthController(
-        ISender sender,
-        SignInManager<ApplicationUser> signInManager)
-    {
-        this.sender = sender;
-        this.signInManager = signInManager;
-    }
-
-   /// <summary>
+    /// <summary>
     /// Initiates Google OAuth login flow.
     /// </summary>
-    /// <response code="302">Redirect to Google OAuth login page.</response>
+    /// <remarks>
+    /// Requires a <b>deviceId</b> which is forwarded through the Google OAuth process
+    /// and returned to the callback route for token generation.
+    /// </remarks>
     [HttpGet("login/google")]
     [EndpointName("LoginWithGoogle")]
     [EndpointSummary("Redirects to Google OAuth login.")]
-    [EndpointDescription("Starts Google OAuth sign-in flow.")]
+    [EndpointDescription("Starts Google OAuth sign-in flow and preserves the deviceId.")]
     [ProducesResponseType(StatusCodes.Status302Found)]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
     [AllowAnonymous]
-    public IActionResult LoginWithGoogle()
+
+    public IActionResult LoginWithGoogle(
+    [FromQuery] string deviceId,
+    [FromQuery] string redirectUrl)
     {
-        string redirectUrl = Url.Action(
-            action: nameof(GoogleCallback),
-            controller: "ExternalAuth",
-            values: null,
+        if (string.IsNullOrWhiteSpace(deviceId))
+            return BadRequest(new { error = "DeviceId is required." });
+
+        if (string.IsNullOrWhiteSpace(redirectUrl))
+            return BadRequest(new { error = "RedirectUrl is required." });
+
+        string callbackUrl = Url.Action(
+            nameof(GoogleCallback),
+            "ExternalAuth",
+            values: new { deviceId, redirectUrl },
             protocol: Request.Scheme
-        ) ?? string.Empty;
+        )!;
 
         AuthenticationProperties props =
             signInManager.ConfigureExternalAuthenticationProperties(
-                provider: "Google",
-                redirectUrl: redirectUrl
-            );
+                GoogleProvider,
+                callbackUrl);
 
-        return Challenge(props, "Google");
+        return Challenge(props, GoogleProvider);
     }
 
     /// <summary>
     /// Google OAuth callback endpoint.
     /// </summary>
     /// <remarks>
-    /// Google returns the authenticated user details here.  
-    /// The endpoint then calls MediatR to produce JWT + refresh tokens.
+    /// Receives Google OAuth response, generates JWT + refresh token (scoped by deviceId),
+    /// and writes auth cookies.
     /// </remarks>
-    /// <response code="200">Successfully authenticated and tokens returned.</response>
-    /// <response code="401">Authentication failed.</response>
-    /// <response code="500">Internal server error.</response>
     [HttpGet("google/callback")]
     [EndpointName("GoogleCallback")]
     [EndpointSummary("Google authentication callback.")]
     [EndpointDescription("Receives Google OAuth response and generates user tokens.")]
-    [ProducesResponseType(typeof(AuthResponse), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status302Found)]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
     [AllowAnonymous]
-    public async Task<ActionResult<AuthResponse>> GoogleCallback(CancellationToken ct)
+    public async Task<IActionResult> GoogleCallback(
+    [FromQuery] string deviceId,
+    [FromQuery] string redirectUrl,
+    CancellationToken ct)
+    {
+        if (string.IsNullOrWhiteSpace(deviceId))
+            return BadRequest("missing_device_id");
+
+        Result<AuthDto> result =
+            await sender.Send(new GoogleLoginCommand(deviceId), ct);
+
+        if (result.IsFailure)
+            return Redirect($"{redirectUrl}?error=google_login_failed");
+
+        authCookieWriter.Write(result.TryGetValue());
+
+        return Redirect(redirectUrl);
+    }
+
+
+    private string BuildErrorRedirect(string error)
+    {
+        string callbackUrl = frontendOptions.AuthCallbackUrl;
+        return $"{callbackUrl}?error={Uri.EscapeDataString(error)}";
+    }
+
+
+
+    [Authorize]
+    [HttpGet("me")]
+    public async Task<ActionResult<AuthResponse>> Me(CancellationToken ct)
     {
         Result<AuthResponse> result =
-            await sender.Send(new GoogleLoginCommand(), ct);
+            await sender.Send(new GetMeQuery(), ct);
 
-        return result.ToActionResult<AuthResponse>(HttpContext);
+        return result.ToActionResult(HttpContext);
     }
 }
