@@ -10,8 +10,9 @@ using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
+using Microsoft.Extensions.Options;
 using Microsoft.OpenApi;
+using quiz_management_system.App.Implemntation;
 using quiz_management_system.Application.Common.Behaivors;
 using quiz_management_system.Application.Common.Settings;
 using quiz_management_system.Application.Interfaces;
@@ -19,10 +20,10 @@ using quiz_management_system.Infrastructure.Data;
 using quiz_management_system.Infrastructure.Data.Interceptors;
 using quiz_management_system.Infrastructure.Email;
 using quiz_management_system.Infrastructure.Idenitity;
+using quiz_management_system.Infrastructure.Services;
+using Resend;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System.Reflection;
-using System.Text;
-using Microsoft.OpenApi.Models;
 
 namespace quiz_management_system.App;
 
@@ -52,7 +53,9 @@ public static class ServiceRegistration
             .ConfigureForwardedHeaders()
             .ConfigureMappings()
             .ConfigureProblems()
-            .AddUserContext();
+            .AddUserContext()
+            .AddCookieWriter()
+            .AddFrontendOptions(configuration);
 
         return services;
     }
@@ -93,81 +96,48 @@ public static class ServiceRegistration
 
 
     private static IServiceCollection AddJwtConfiguration(
-        this IServiceCollection services,
-        IConfiguration configuration)
+     this IServiceCollection services,
+     IConfiguration configuration)
     {
-        services.AddOptions<JwtSettings>()
+        services
+            .AddOptions<JwtSettings>()
             .BindConfiguration(JwtSettings.SectionName)
             .ValidateDataAnnotations()
             .ValidateOnStart();
 
-        var jwtSettings = configuration
-            .GetSection(JwtSettings.SectionName)
-            .Get<JwtSettings>()
-            ?? throw new InvalidOperationException(
-                $"{JwtSettings.SectionName} section missing");
-        services.AddSingleton(jwtSettings);
+        services.AddSingleton(sp =>
+            sp.GetRequiredService<IOptions<JwtSettings>>().Value);
 
         services
             .AddAuthentication(options =>
             {
-
                 options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
                 options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-
             })
-            .AddJwtBearer(JwtBearerDefaults.AuthenticationScheme, options =>
+            .AddJwtBearer();
+
+        services.ConfigureOptions<JwtBearerOptionsConfigurator>();
+
+        services.AddAuthentication()
+            .AddGoogle(options =>
             {
-                options.TokenValidationParameters = new TokenValidationParameters
-                {
-                    ValidateIssuer = true,
-                    ValidIssuer = jwtSettings.Issuer,
-                    ValidateAudience = true,
-                    ValidAudience = jwtSettings.Audience,
-                    ValidateIssuerSigningKey = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(
-                    Encoding.UTF8.GetBytes(jwtSettings.SecretKey)),
-                    ValidateLifetime = true,
-                    ClockSkew = TimeSpan.Zero
-                };
+                options.ClientId = configuration["Authentication:Google:ClientId"]
+                    ?? throw new InvalidOperationException("Google ClientId is missing");
 
-                options.Events = new JwtBearerEvents
-                {
-                    OnMessageReceived = context =>
-                    {
-                        if (context.Request.Cookies.TryGetValue("access_token", out var token))
-                        {
-                            context.Token = token;
-                        }
-                        return Task.CompletedTask;
+                options.ClientSecret = configuration["Authentication:Google:ClientSecret"]
+                    ?? throw new InvalidOperationException("Google ClientSecret is missing");
 
-                    }
-                };
+                options.CallbackPath = "/signin-google";
+                options.SignInScheme = IdentityConstants.ExternalScheme;
+                options.SaveTokens = true;
 
-            })
-        .AddGoogle(options =>
-        {
-            options.ClientId = configuration["Authentication:Google:ClientId"]
-                ?? throw new InvalidOperationException("Google ClientId is missing");
-
-            options.ClientSecret = configuration["Authentication:Google:ClientSecret"]
-                ?? throw new InvalidOperationException("Google ClientSecret is missing");
-
-              options.CallbackPath = "/signin-google";
-            options.SignInScheme = IdentityConstants.ExternalScheme;
-
-            options.SaveTokens = true;
-
-            options.Scope.Add("email");
-            options.Scope.Add("profile");
-
-
-
-
-        });
+                options.Scope.Add("email");
+                options.Scope.Add("profile");
+            });
 
         return services;
     }
+
 
     private static IServiceCollection AddControllersWithVersioning(this IServiceCollection services)
     {
@@ -258,7 +228,7 @@ public static class ServiceRegistration
         {
             options.AddPolicy("AllowFrontend", policy =>
                 policy
-                    .WithOrigins("http://127.0.0.1:5500")
+                    .WithOrigins("http://localhost:3000")
                     .AllowAnyHeader()
                     .AllowAnyMethod()
                     .AllowCredentials());
@@ -277,6 +247,7 @@ public static class ServiceRegistration
         services.AddScoped<UpdatableEntityInterceptor>();
         services.AddScoped<CreatableEntityInterceptor>();
 
+        services.AddScoped<SoftDeleteEntityInterceptor>();
 
         //var connectionString = configuration.GetConnectionString("DefaultConnection");
 
@@ -290,7 +261,9 @@ public static class ServiceRegistration
 
         .AddInterceptors(
         sp.GetRequiredService<CreatableEntityInterceptor>(),
-        sp.GetRequiredService<UpdatableEntityInterceptor>())
+        sp.GetRequiredService<UpdatableEntityInterceptor>(),
+        sp.GetRequiredService<SoftDeleteEntityInterceptor>())
+
         );
 
 
@@ -314,40 +287,40 @@ public static class ServiceRegistration
     }
 
 
-    public static IServiceCollection AddMessageSending(
-     this IServiceCollection services,
-        IConfiguration configuration)
-    {
-        services.Configure<EmailSettings>(configuration.GetSection("EmailSettings"));
-
-
-        services.AddScoped<IEmailSender, EmailSender>();
-
-
-        return services;
-
-    }
-
     //public static IServiceCollection AddMessageSending(
-    //     this IServiceCollection services,
-    //     IConfiguration configuration)
+    // this IServiceCollection services,
+    //    IConfiguration configuration)
     //{
-    //    services.Configure<ResendSettings>(configuration.GetSection("Resend"));
+    //    services.Configure<EmailSettings>(configuration.GetSection("EmailSettings"));
 
-    //    services.AddOptions();
-    //    services.AddHttpClient<ResendClient>();
 
-    //    services.Configure<ResendClientOptions>(o =>
-    //    {
-    //        o.ApiToken = configuration["Resend:ApiKey"]!;
-    //    });
+    //    services.AddScoped<IEmailSender, EmailSender>();
 
-    //    services.AddTransient<IResend, ResendClient>();
-
-    //    services.AddScoped<IEmailSender, ResendEmailSender>();
 
     //    return services;
+
     //}
+
+    public static IServiceCollection AddMessageSending(
+         this IServiceCollection services,
+         IConfiguration configuration)
+    {
+        services.Configure<ResendSettings>(configuration.GetSection("Resend"));
+
+        services.AddOptions();
+        services.AddHttpClient<ResendClient>();
+
+        services.Configure<ResendClientOptions>(o =>
+        {
+            o.ApiToken = configuration["Resend:ApiKey"]!;
+        });
+
+        services.AddTransient<IResend, ResendClient>();
+
+        services.AddScoped<IEmailSender, ResendEmailSender>();
+
+        return services;
+    }
 
     private static IServiceCollection ConfigureBackGroundJobs(this IServiceCollection services, IConfiguration configuration)
     {
@@ -415,6 +388,26 @@ public static class ServiceRegistration
         services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
         services.AddScoped<IUserContext, HttpUserContext>();
 
+
+        return services;
+    }
+    private static IServiceCollection AddCookieWriter(this IServiceCollection services)
+    {
+        services.AddHttpContextAccessor();
+        services.AddScoped<IAuthCookieWriter, AuthCookieWriter>();
+        return services;
+    }
+
+
+    public static IServiceCollection AddFrontendOptions(
+         this IServiceCollection services,
+         IConfiguration configuration)
+    {
+        services.Configure<FrontendOptions>(
+            configuration.GetSection("Frontend"));
+
+        services.AddSingleton(sp =>
+            sp.GetRequiredService<IOptions<FrontendOptions>>().Value);
 
         return services;
     }
