@@ -7,7 +7,10 @@ namespace quiz_management_system.Application.Features.Courses.Queries.GetAllCour
 public sealed record GetAllCoursesQuery()
     : ICachedQuery<Result<IReadOnlyList<CourseResponse>>>
 {
-    public string CacheKey => "courses:all";
+
+    public static string GetCacheKey() => "courses:all";
+
+    public string CacheKey => GetCacheKey();
     public string[] Tags => new[] { "courses" };
     public TimeSpan Expiration => TimeSpan.FromHours(24);
 }
