@@ -17,9 +17,16 @@ public sealed class CreateAdminHandler(
         CancellationToken ct)
     {
 
+
+        string username = string.Concat(
+    request.FullName
+        .Trim()
+        .Split(' ', StringSplitOptions.RemoveEmptyEntries)
+);
+
         var registrationResult = await identityService.CreateIdentityByEmailAsync(
             request.Email,
-            request.FullName,
+            username,
             DefaultRoles.Admin,
             ct
         );
