@@ -1,5 +1,4 @@
-﻿using Mapster;
-using MediatR;
+﻿using MediatR;
 using quiz_management_system.Application.Dtos;
 using quiz_management_system.Application.Interfaces;
 using quiz_management_system.Contracts.Reponses.Identity;
@@ -7,23 +6,34 @@ using quiz_management_system.Domain.Common.ResultPattern.Result;
 
 namespace quiz_management_system.Application.Features.Login;
 
-public sealed class LoginByEmailCommandHandler(IIdentityService identityService, ITokenProvider tokenProvider) : IRequestHandler<LoginCommand, Result<AuthResponse>>
+public sealed class LoginCommandHandler(
+    IIdentityService identityService,
+    ITokenProvider tokenProvider
+) : IRequestHandler<LoginCommand, Result<AuthDto>>
 {
-
-    public async Task<Result<AuthResponse>> Handle(LoginCommand request, CancellationToken cancellationToken)
+    public async Task<Result<AuthDto>> Handle(
+        LoginCommand request,
+        CancellationToken cancellationToken)
     {
+        Result<AuthenticatedUser> authResult =
+            await identityService.AuthenticateByEmailAsync(
+                request.Email,
+                request.Password);
 
-
-        Result<AuthenticatedUser> authResult = await identityService.AuthenticateByEmailAsync(request.Email, request.Password);
         if (authResult.IsFailure)
-            return Result.Failure<AuthResponse>(authResult.TryGetError());
+            return Result.Failure<AuthDto>(authResult.TryGetError());
 
-        AuthenticatedUser authUser = authResult.TryGetValue();
+        AuthenticatedUser user = authResult.TryGetValue();
 
-        Result tokenResult = await tokenProvider.GenerateJwtTokenAsync(authUser, cancellationToken);
+        Result<AuthDto> tokenResult =
+            await tokenProvider.GenerateJwtTokenAsync(
+                user,
+                request.DeviceId,
+                cancellationToken);
+
         if (tokenResult.IsFailure)
-            return Result.Failure<AuthResponse>(tokenResult.TryGetError());
+            return Result.Failure<AuthDto>(tokenResult.TryGetError());
 
-        return Result.Success(authUser.Adapt<AuthResponse>());
+        return tokenResult;
     }
 }
