@@ -11,14 +11,14 @@ namespace quiz_management_system.Domain.Users.Abstraction;
 
 
 
-public abstract class DomainUser : AggregateRoot, IAuditable
+public abstract class DomainUser : AggregateRoot, IAuditable, ISoftDeletable
 {
     public string? PictureUrl { get; private set; } = string.Empty;
 
     public string FullName { get; protected set; } = string.Empty;
     public string Email { get; protected set; } = string.Empty;
 
-    public StudentStatus Status { get; private set; } = StudentStatus.Active;
+    public UserStatus Status { get; private set; } = UserStatus.Active;
     public Role Role { get; private set; }
 
 
@@ -29,6 +29,12 @@ public abstract class DomainUser : AggregateRoot, IAuditable
     public Guid CreatedBy { get; protected set; } = Guid.Empty;
     public DateTimeOffset LastModifiedUtc { get; protected set; } = DateTimeOffset.UtcNow;
     public Guid LastModifiedBy { get; protected set; } = Guid.Empty;
+
+
+    public bool IsDeleted { get; protected set; }
+    public Guid? DeletedById { get; protected set; } = Guid.Empty;
+    public DateTimeOffset? DeletedOn { get; protected set; }
+
 
 
     DateTimeOffset ICreatable.CreatedAtUtc
@@ -54,6 +60,26 @@ public abstract class DomainUser : AggregateRoot, IAuditable
         get => LastModifiedBy;
         set => LastModifiedBy = value;
     }
+
+
+    bool ISoftDeletable.IsDeleted
+    {
+        get => IsDeleted;
+        set => IsDeleted = value;
+    }
+
+    Guid? ISoftDeletable.DeletedById
+    {
+        get => DeletedById;
+        set => DeletedById = value;
+    }
+
+    DateTimeOffset? ISoftDeletable.DeletedOn
+    {
+        get => DeletedOn;
+        set => DeletedOn = value;
+    }
+
 
     protected DomainUser() { }
 
@@ -94,14 +120,57 @@ public abstract class DomainUser : AggregateRoot, IAuditable
     }
     public Result ActivateUser()
     {
-        Status = StudentStatus.Active;
+        Status = UserStatus.Active;
         return Result.Success();
     }
 
     public Result DisActivateUser()
     {
-        Status = StudentStatus.InActive;
+        Status = UserStatus.InActive;
         return Result.Success();
     }
+
+
+
+    public Result Restore()
+    {
+        if (!IsDeleted)
+        {
+            return Result.Failure(
+                DomainError.InvalidState(nameof(DomainUser), "User is not deleted."));
+        }
+
+        IsDeleted = false;
+        DeletedById = Guid.Empty;
+        DeletedOn = null;
+        Status = UserStatus.Active;
+
+
+
+
+        return Result.Success();
+    }
+
+    public Result SoftDelete(Guid deletedBy)
+    {
+        if (IsDeleted)
+        {
+            return Result.Failure(
+                DomainError.InvalidState(nameof(DomainUser), "User is already deleted."));
+        }
+
+        IsDeleted = true;
+        DeletedById = deletedBy;
+        DeletedOn = DateTimeOffset.UtcNow;
+        Status = UserStatus.InActive;
+
+        AddDomainEvent(
+            new UserDeletedEvent(Id, DeletedById.Value, DeletedOn.Value));
+
+        return Result.Success();
+    }
+
+
+
 
 }
